refactor(ui): extract react flow edges builder
