+++ tests/core/cleanup/noLegacySwimlaneArtifacts.test.ts
@@ -156,6 +156,13 @@ describe('legacy swimlane cleanup', () => {
     expect(gridCanvas.includes('const createFlowNode = (')).toBe(false);
   });
 
+  it('extracts react flow edges builder into a dedicated helper', () => {
+    expect(existsSync(pathInRepo('src/ui/components/Canvas/buildReactFlowEdges.ts'))).toBe(true);
+
+    const gridCanvas = readFileSync(pathInRepo('src/ui/components/Canvas/GridCanvas.tsx'), 'utf8');
+    expect(gridCanvas.includes('const HANDLE_MAP')).toBe(false);
+  });
+
   it('groups vertical slice aggregate files under a dedicated domain subfolder', () => {
     expect(existsSync(pathInRepo('src/core/domain/VerticalSlice.ts'))).toBe(false);
     expect(existsSync(pathInRepo('src/core/domain/VerticalSliceCollection.ts'))).toBe(false);
